feat: add bot menu item, move from keys scene
